refactor(s4): isolate pure flywheel state reducer
